# Conflicts:
#	.github/scripts/rust/Makefile.toml

+++ packages/graph/hash_graph/bench/benches/util.rs
@@ -1,10 +1,11 @@
-use std::{mem::ManuallyDrop, str::FromStr};
+use std::str::FromStr;
 
 use graph::{
     ontology::AccountId,
     store::{
-        AsClient, DataTypeStore, DatabaseConnectionInfo, DatabaseType, EntityTypeStore,
-        LinkTypeStore, PostgresStorePool, PropertyTypeStore, StorePool,
+        AsClient, BaseUriAlreadyExists, DataTypeStore, DatabaseConnectionInfo, DatabaseType,
+        EntityTypeStore, LinkTypeStore, PostgresStore, PostgresStorePool, PropertyTypeStore,
+        StorePool,
     },
 };
 use tokio::runtime::Runtime;
@@ -16,14 +17,13 @@ pub type Store = <Pool as StorePool>::Store<'static>;
 
 // TODO - deduplicate with integration/postgres/mod.rs
 pub struct StoreWrapper {
-    bench_db_name: String,
-    source_db_pool: Pool,
-    pool: ManuallyDrop<Pool>,
-    pub store: ManuallyDrop<Store>,
+    pub bench_db_name: String,
+    _pool: Pool,
+    pub store: Store,
 }
 
 impl StoreWrapper {
-    pub async fn new(bench_db_name: &str) -> Self {
+    pub async fn new(bench_db_name: &str, fail_on_exists: bool) -> Self {
         let source_db_connection_info = DatabaseConnectionInfo::new(
             // TODO - get these from env
             //  https://app.asana.com/0/0/1203071961523005/f
@@ -56,32 +56,53 @@ impl StoreWrapper {
                 .expect("could not acquire a database connection");
             let client = conn.as_client();
 
-            client
-                .execute(
+            let exists: bool = client
+                .query_one(
                     r#"
-                    /* KILL ALL EXISTING CONNECTION FROM ORIGINAL DB*/
-                    SELECT pg_terminate_backend(pg_stat_activity.pid) FROM pg_stat_activity
-                    WHERE pg_stat_activity.datname = $1 AND pid <> pg_backend_pid();
+                    SELECT EXISTS(
+                        SELECT 1 FROM pg_catalog.pg_database WHERE datname = $1
+                    );
                     "#,
-                    &[&source_db_connection_info.database()],
+                    &[&bench_db_name],
                 )
                 .await
-                .expect("failed to kill existing connections");
+                .expect("failed to check if database exists")
+                .get(0);
 
-            client
-                .execute(
-                    &format!(
+            assert!(
+                !(fail_on_exists && exists),
+                "database `{}` exists, and `fails_on_exists` was set to true",
+                bench_db_name
+            );
+
+            if !(exists) {
+                client
+                    .execute(
                         r#"
-                        /* CLONE DATABASE TO NEW ONE */
-                        CREATE DATABASE {bench_db_name} WITH TEMPLATE {} OWNER {};
+                        /* KILL ALL EXISTING CONNECTION FROM ORIGINAL DB*/
+                        SELECT pg_terminate_backend(pg_stat_activity.pid) FROM pg_stat_activity
+                        WHERE pg_stat_activity.datname = $1 AND pid <> pg_backend_pid();
                         "#,
-                        source_db_connection_info.database(),
-                        bench_db_connection_info.user()
-                    ),
-                    &[],
-                )
-                .await
-                .expect("failed to clone database");
+                        &[&source_db_connection_info.database()],
+                    )
+                    .await
+                    .expect("failed to kill existing connections");
+
+                client
+                    .execute(
+                        &format!(
+                            r#"
+                            /* CLONE DATABASE TO NEW ONE */
+                            CREATE DATABASE {bench_db_name} WITH TEMPLATE {} OWNER {};
+                            "#,
+                            source_db_connection_info.database(),
+                            bench_db_connection_info.user()
+                        ),
+                        &[],
+                    )
+                    .await
+                    .expect("failed to clone database");
+            }
         }
 
         let pool = PostgresStorePool::new(&bench_db_connection_info, NoTls)
@@ -95,106 +116,112 @@ impl StoreWrapper {
             .expect("could not acquire a database connection");
 
         Self {
-            source_db_pool,
             bench_db_name: bench_db_name.to_owned(),
-            pool: ManuallyDrop::new(pool),
-            store: ManuallyDrop::new(store),
-        }
-    }
-
-    pub async fn seed<D, P, L, E>(
-        &mut self,
-        account_id: AccountId,
-        data_types: D,
-        property_types: P,
-        link_types: L,
-        entity_types: E,
-    ) where
-        D: IntoIterator<Item = &'static str>,
-        P: IntoIterator<Item = &'static str>,
-        L: IntoIterator<Item = &'static str>,
-        E: IntoIterator<Item = &'static str>,
-    {
-        for data_type in data_types {
-            self.store
-                .create_data_type(
-                    DataType::from_str(data_type).expect("could not parse data type"),
-                    account_id,
-                )
-                .await
-                .expect("failed to create data type");
-        }
-
-        for property_type in property_types {
-            self.store
-                .create_property_type(
-                    PropertyType::from_str(property_type).expect("could not parse property type"),
-                    account_id,
-                )
-                .await
-                .expect("failed to create property type");
-        }
-
-        // Insert link types before entity types so entity types can refer to them
-        for link_type in link_types {
-            self.store
-                .create_link_type(
-                    LinkType::from_str(link_type).expect("could not parse link type"),
-                    account_id,
-                )
-                .await
-                .expect("failed to create link type");
-        }
-
-        for entity_type in entity_types {
-            self.store
-                .create_entity_type(
-                    EntityType::from_str(entity_type).expect("could not parse entity type"),
-                    account_id,
-                )
-                .await
-                .expect("failed to create entity type");
+            _pool: pool,
+            store,
         }
     }
 }
 
-impl Drop for StoreWrapper {
-    fn drop(&mut self) {
-        // We're in the process of dropping the parent struct, we just need to ensure we release
-        // the connections of this pool before deleting the database
-        // SAFETY: The values of `store` and `pool` are not accessed after dropping
-        unsafe {
-            ManuallyDrop::drop(&mut self.store);
-            ManuallyDrop::drop(&mut self.pool);
+pub async fn seed<D, P, L, E, C>(
+    store: &mut PostgresStore<C>,
+    account_id: AccountId,
+    data_types: D,
+    property_types: P,
+    link_types: L,
+    entity_types: E,
+) where
+    D: IntoIterator<Item = &'static str>,
+    P: IntoIterator<Item = &'static str>,
+    L: IntoIterator<Item = &'static str>,
+    E: IntoIterator<Item = &'static str>,
+    C: AsClient,
+{
+    for data_type_str in data_types {
+        let data_type = DataType::from_str(data_type_str).expect("could not parse data type");
+
+        match store.create_data_type(data_type.clone(), account_id).await {
+            Ok(_) => {}
+            Err(report) => {
+                if report.contains::<BaseUriAlreadyExists>() {
+                    store
+                        .update_data_type(data_type, account_id)
+                        .await
+                        .expect("failed to update data type");
+                } else {
+                    Err(report).expect("failed to create data type")
+                }
+            }
         }
+    }
 
-        let runtime = Runtime::new().unwrap();
-        runtime.block_on(async {
-            let conn = self
-                .source_db_pool
-                .acquire_owned()
-                .await
-                .expect("could not acquire a database connection");
+    for property_type_str in property_types {
+        let property_type =
+            PropertyType::from_str(property_type_str).expect("could not parse property type");
 
-            conn.as_client()
-                .execute(
-                    &format!(
-                        r#"
-                        DROP DATABASE IF EXISTS {};
-                        "#,
-                        self.bench_db_name
-                    ),
-                    &[],
-                )
-                .await
-                .expect("failed to drop database");
-        });
+        match store
+            .create_property_type(property_type.clone(), account_id)
+            .await
+        {
+            Ok(_) => {}
+            Err(report) => {
+                if report.contains::<BaseUriAlreadyExists>() {
+                    store
+                        .update_property_type(property_type, account_id)
+                        .await
+                        .expect("failed to update property type");
+                } else {
+                    Err(report).expect("failed to create property type")
+                }
+            }
+        }
+    }
+
+    // Insert link types before entity types so entity types can refer to them
+    for link_type_str in link_types {
+        let link_type = LinkType::from_str(link_type_str).expect("could not parse link type");
+
+        match store.create_link_type(link_type.clone(), account_id).await {
+            Ok(_) => {}
+            Err(report) => {
+                if report.contains::<BaseUriAlreadyExists>() {
+                    store
+                        .update_link_type(link_type, account_id)
+                        .await
+                        .expect("failed to update link type");
+                } else {
+                    Err(report).expect("failed to create link type")
+                }
+            }
+        }
+    }
+
+    for entity_type_str in entity_types {
+        let entity_type =
+            EntityType::from_str(entity_type_str).expect("could not parse entity type");
+
+        match store
+            .create_entity_type(entity_type.clone(), account_id)
+            .await
+        {
+            Ok(_) => {}
+            Err(report) => {
+                if report.contains::<BaseUriAlreadyExists>() {
+                    store
+                        .update_entity_type(entity_type, account_id)
+                        .await
+                        .expect("failed to update entity type");
+                } else {
+                    Err(report).expect("failed to create entity type")
+                }
+            }
+        }
     }
 }
 
-pub fn setup(db_name: &str) -> (Runtime, StoreWrapper) {
+pub fn setup(db_name: &str, fail_on_exists: bool) -> (Runtime, StoreWrapper) {
     let runtime = Runtime::new().unwrap();
 
-    let store_wrapper = runtime.block_on(StoreWrapper::new(db_name));
+    let store_wrapper = runtime.block_on(StoreWrapper::new(db_name, fail_on_exists));
     (runtime, store_wrapper)
 }
